rename: 컨트롤러 클래스명 변경 및 메서드 이동
- RequestMapping으로 묶을 수 있는 엔드포인트를 기준
