+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,37 +1,28 @@
 package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.request.ReservationRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
 import java.net.URI;
-import java.util.List;
 
 @RestController
-public class AdminReservationController {
+@RequestMapping("/admin")
+public class AdminController {
     private final ReservationService reservationService;
 
-    public AdminReservationController(final ReservationService reservationService) {
+    public AdminController(final ReservationService reservationService) {
         this.reservationService = reservationService;
     }
 
-    @PostMapping("/admin/reservations")
+    @PostMapping("/reservations")
     public ResponseEntity<ReservationResponse> createAdminReservation(@RequestBody final ReservationRequest request) {
         ReservationResponse reservationResponse = reservationService.create(request);
         return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id())).body(reservationResponse);
-    }
-
-    @GetMapping("/reservations")
-    public ResponseEntity<List<ReservationResponse>> readReservations() {
-        List<ReservationResponse> reservations = reservationService.findEntireReservationList();
-        return ResponseEntity.ok(reservations);
-    }
-
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> delete(@PathVariable final Long id) {
-        reservationService.delete(id);
-        return ResponseEntity.noContent().build();
     }
 }
