+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,0 +1,54 @@
+package roomescape.controller;
+
+import jakarta.validation.Valid;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.dto.LoginMember;
+import roomescape.dto.request.MemberReservationRequest;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.service.ReservationService;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+@RequestMapping("/reservations")
+public class ReservationController {
+    private final ReservationService reservationService;
+
+    public ReservationController(final ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @PostMapping()
+    public ResponseEntity<ReservationResponse> createReservationByClient(@Valid @RequestBody final MemberReservationRequest memberRequest,
+                                                                         final LoginMember member) {
+        ReservationRequest reservationRequest = new ReservationRequest(member.id(), memberRequest.date(),
+                memberRequest.timeId(), memberRequest.themeId());
+        ReservationResponse reservationResponse = reservationService.create(reservationRequest);
+        return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id())).body(reservationResponse);
+    }
+
+    @GetMapping()
+    public ResponseEntity<List<ReservationResponse>> readReservations() {
+        List<ReservationResponse> reservations = reservationService.findEntireReservationList();
+        return ResponseEntity.ok(reservations);
+    }
+
+    @GetMapping(params = {"memberId", "themeId", "dateFrom", "dateTo"})
+    public ResponseEntity<List<ReservationResponse>> findAllByMemberAndThemeAndPeriod(
+            @RequestParam(required = false, name = "memberId") final Long memberId,
+            @RequestParam(required = false, name = "themeId") final Long themeId,
+            @RequestParam(required = false, name = "dateFrom") final String dateFrom,
+            @RequestParam(required = false, name = "dateTo") final String dateTo) {
+        return ResponseEntity.ok(
+                reservationService.findDistinctReservations(memberId, themeId, dateFrom, dateTo));
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable final Long id) {
+        reservationService.delete(id);
+        return ResponseEntity.noContent().build();
+    }
+}
